Mover `classifyRouteCapture` to separate file
